Fixed ViewPager2 setCurrentItem not working bug
Before, after going to a SignDetailFragment then going back to the SignListFragment, the ViewPager2's setCurrentItem method wasn't working for the positions on screen. It is now fixed. I'm not sure how, but I think moving the viewPager reference within SignItemFragment helped. It is now a local variable, which probably prevents loading-state errors.